+++ Examples/BasicGame/include/BasicGame.hpp
@@ -1,5 +1,8 @@
 #pragma once
 #include <Pyramid/Core/Game.hpp>
+#include <Pyramid/Graphics/Shader/Shader.hpp>
+#include <Pyramid/Graphics/Buffer/VertexArray.hpp>
+#include <memory> // For std::shared_ptr
 
 class BasicGame : public Pyramid::Game
 {
@@ -11,4 +14,8 @@ protected:
     void onCreate() override;
     void onUpdate(float deltaTime) override;
     void onRender() override;
+
+private:
+    std::shared_ptr<Pyramid::IShader> m_shader;
+    std::shared_ptr<Pyramid::IVertexArray> m_vertexArray;
 };